processing/leaf_segmentation/weyler_phenobench_maskrcnn/requirements.txt
absl-py==2.1.0
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
filelock==3.15.1
fonttools==4.53.0
fsspec==2024.6.0
grpcio==1.64.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2==3.1.4
kiwisolver==1.4.5
lightning-utilities==0.11.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.82
packaging==24.1
pillow==10.3.0
protobuf==4.25.3
pycocotools==2.0.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
six==1.16.0
sympy==1.12.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
torch==2.3.1
torchmetrics==1.4.0.post0
torchvision==0.18.1
tqdm==4.66.4
triton==2.3.1
typing_extensions==4.12.2
Werkzeug==3.0.3
zipp==3.19.2
